+++ src/main/php/lang/ast/Emitter.class.php
@@ -725,6 +725,19 @@ abstract class Emitter {
     $this->emit($instance->member);
   }
 
+  protected function emitDynamicInstance($instance) {
+    if ('new' === $instance->expression->kind) {
+      $this->out->write('(');
+      $this->emit($instance->expression);
+      $this->out->write(')->{');
+    } else {
+      $this->emit($instance->expression);
+      $this->out->write('->{');
+    }
+    $this->emit($instance->member);
+    $this->out->write('}');
+  }
+
   protected function emitNullSafeInstance($instance) {
     $t= $this->temp();
     $this->out->write('null === ('.$t.'= ');